+++ packages/account-eip6963/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sophon-labs/account-eip6963
 
+## 2.0.10-next.5
+
+### Patch Changes
+
+- Fixing provider window error on react native
+- Updated dependencies
+  - @sophon-labs/account-provider@2.0.10-next.5
+  - @sophon-labs/account-core@2.0.10-next.5
+
 ## 2.0.10-next.4
 
 ### Patch Changes